+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Unreleased
 
-* must exist /target folder.
+* inline `def`.
+
+## 0.2.1 (2025-03-31)
+
+* added - org.slf4j/slf4j-simple  {:mvn/version "2.0.17"}
+* warn if `target/db.sqlite` does not exist.
 
 ## 0.2.0 (2025-03-25)